requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
PyJWT==2.8.0
Werkzeug==2.3.7
Flask-CORS==4.0.0
Faker==18.11.2
pytest==7.4.0
python-dotenv==1.0.1
requests==2.32.3
pymysql==1.1.1
pyrofex
